Changes to the interpolator routine for the magnetic field map.
Support for writing and reading text and binary files with all
information stored inside the file (specification is trivial).
This stores the size of the grid along with the data points.
The binary file format should work on 64-bit (untested).

It incorporates arrays, which makes it easier to put the five
components of the magnetic field in one array.  The weights in
the interpolation only have to be calculated once then.  A new
interpolator can be created by using QwInterpolator<float,5>(3).
This will use values of 5 floats on a 3 dimensional grid.  When
only single values are used (the default: QwInterpoltaor<float>)
you can use a simpler syntax for getting interpolated values.

Moved some of the unsafe (direct value read/write) methods to
private.

(This already works as a drop-in replacement in the magnetic
field class, removing low-level functions.  See commit later.)


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.restruct-Q2-det@841 b9700a98-9937-0410-b777-f318d6eb5bff