Merged the full revision range (759:842) from the "pking.intersubsystem_communication" branch into the trunk.
This represents a functioning system for communication
between subsystems.  For an example of use see the
QwMainCerenkovDetector::ExchangeProcessedData
function which pulls beamline parameters into the
main detector subsytem.
In the current version, nothing is done with the beamline
data once it is pulled into the main detector subsystem.

NOTE:  It is now possible for subsystems to "publish" data
       for access by other subsystems; however, please talk
       to Paul or Wouter before you decide that your subsystem
       needs to publish a data element.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@843 b9700a98-9937-0410-b777-f318d6eb5bff