Changes to the reconstruction of the Q^2 in the event (restructuring).
The magnetic field uses the QwInterpolator.  The field is still read from
the old map file (peiqing_2007.dat), but could easily write out a new map
file which has the rotation and rescaling included, and is 5 times smaller
even when also r and phi components are included (gzip doesn't do better).

The QwRayTracer uses the (slightly modified) magnetic field interface.

The front and back partial tracks are passed directly to the Bridge()
function.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.restruct-Q2-det@846 b9700a98-9937-0410-b777-f318d6eb5bff