Updated QwDatabase to provide functions for retrieving main_detector_id (lumi_detector_id) for a given main (lumi) detector.  Same functionality as GetMonitorID(string).  Maps a detector name onto an integer using an associative array.
Important Note:  The associative arrays for storing main detector/lumi detector/beam monitor IDs are implemented as std::map<string, unsigned int>.  Each is declared as a static member of the QwDatabase class.  There will be issues with this implementation choice if we ever try to use the QwDatabase class to access two different databases simultaneously.  (The ID for a given detector may not be the same between database instances.)



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/spayde.db.00_01_0000@925 b9700a98-9937-0410-b777-f318d6eb5bff