Two major updates to the Parity Analyzer that will affect all the subsystems.
	1. Command Line Options Decoding
	2. Updated event cut framework



=Command Line Options Decoding=
When command line options are added for a subsystem, they can be accessed using the 
	VQwSubsystemParity::ProcessOptions(QwOptions &)
routine. At the moment, for Parity subsystems, command line options are set in the 
	QwOptionsParity::DefineOptionsParity(QwOptions &)
routine, which is called in the QwAnalysis main routine. For an example see QwHelicity subsystem.


=Updated event cut framework=
Now event cuts can be set to individual channels in a device. For example in BPMStriplines cuts can be set to relative X and Y, absolute X and Y, individual wires (XP, XM, YP, YM) or four wire sum.

==Following routines are added/modified==

This sets the upper and lower limits (fULimit and fLLimit) for the channel.
	void QwVQWK_Channel::SetSingleEventCuts(Double_t min, Double_t max);


Check upper and lower limits (fULimit and fLLimit) set on this channel 
	Bool_t QwVQWK_Channel::ApplySingleEventCuts();

The old routine that used to check even cut limits is still exists, but it will only return true/false and will not update even cut related flags or counters.

To set the limit for the given channel of a BPMStripline. 
	void QwBPMStripline::SetSingleEventCuts(TString ch_name, Double_t minX, Double_t maxX);
Where channel name is any of relx, rely, absx, absy, wsum, xp, xm, yp or ym

Added two routines to retrieve upper and lower limit of a VQWK_Channel
	Double_t GetEventCutUpperLimit()
	Double_t GetEventCutLowerLimit()


== In the QwBPMStripline class ==

1. ApplySingleEventCuts() routine, QwVQWK_Channel::ApplySingleEventCuts() is called for Rel X & Y, Abs X & Y, 4 wires and WSum VQWK_Channel objects.

2. SetSingleEventCuts(TString ch_name, Double_t minX, Double_t maxX) routine calls all the 
VQWK_Channel::SetSingleEventCuts(TString ch_name, Double_t minX, Double_t maxX) routines.


The map file is updated so that limit are set to individual channels and to take extra parameter to identify which channel. For example 

	!device_type, device_name, channel_name, lower limit, upper limit
	bpmstripline, qwk_qpd, relx, 2.67319, 	4.45531
	bpmstripline, qwk_qpd, rely, 1.31491, 	2.19152



Rakitha	



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@984 b9700a98-9937-0410-b777-f318d6eb5bff