Correlation Plots -------------------
Created a macro and an exe that can plot the correlations between main detector parameters and the beam parameters.
The macro, correlations.cc, can plot any two parameters that are defined in the HEL Tree in a root file.
The exe, correlation_plots, can plot either he left PMT or the right PMT parameters (yields/asyms) with any of the parameters in the HEL Tree. Since there are no data available for the PMTs yet, I used bpm data from injector runs ( e.e PREX 1672) to test these two files. So the PMT arrays in the codes use the ames of the bpms right now. Later when we have data and the names of the PMTS we have to change these arrays accordingly.

The event cuts are based on cleandata and DeviceErrorCodes that are generated for each device. If anyone wants to be able to apply event cuts to look at specific ranges they have to modify the event cuts or ask me how to do it. 

The macro saves the canvas as a .gif file that contains the run number and devices that were plotted.
The exe prints two canvases as .gifs. One of them contains the correlations plots for all the octents and the other contains a variation of the slope and the offset of the linear fits of the correlations for the octents. The fit parameters are also saved in to a .txt file with the number of entries and the reduced chi square values.  

Please refer the .cc files of the macros and exe for instructions on how to use and compile them.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1001 b9700a98-9937-0410-b777-f318d6eb5bff