Create a branch for development of the database version 01.00.0000.
This branch includes support for a runlet-based analysis.  Gives the option of writing to the database at the end of each file segment rather than at the end of each run (a finer time increment).

Also incorporates support for database schema checking.  Code will exit if attempt to connect to a database schema version inconsistent with QwAnalsyis version.

This branch is not compatible with pre-01.00.0000 databases.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/spayde.db.01_00_0000@1039 b9700a98-9937-0410-b777-f318d6eb5bff