The real time analyzer is setup. It is a light weight QwAnalyzer without the TTree support so the program runs at a reasonable speed to accommodate real time CODA buffer. It is QwAnalysis_Online.cc
Also TFile is replaced with a TMapFile, a memory mapped root file that can be accessed by a GUI application while it is writing. 

The file name is QwMemMapFile.map and currently it is default.

GNUmakefile is modified to create the qwanalysis_online executable.

Known issues,
At the moment there is a run time error at the end of the program run,

*** Break *** segmentation violation
Segmentation fault

The memory sharing has to be done properly to prevent this. This will be addressed in the future.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1047 b9700a98-9937-0410-b777-f318d6eb5bff