Note: this is the wdconinc.running_sum branch.
Changes to implement the running sum.  This now works as
expected.  Numbers agree with ROOT on both asymmetry and
yield average for patterns and single events.  Block and
hardware sum average, variance and sigma are calculated.

The reference for the algorithm is
 * Alternative algorithms (including for higher order moments) are supplied in
 *  Pebay, Philippe (2008), "Formulas for Robust, One-Pass Parallel Computation
 *  of Covariances and Arbitrary-Order Statistical Moments", Technical Report
 *  SAND2008-6212, Sandia National Laboratories.
 *  http://infoserve.sandia.gov/sand_doc/2008/086212.pdf

Sorry, no more time to explain this, gotta run ;-)



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.running_sum@1072 b9700a98-9937-0410-b777-f318d6eb5bff