Finally figured out a way to get rid of the slew of error messages when you run qwsimtracking on a Geant4 ROOT file.  Because of a limitation in the dependency checking(*) this change will not take any effect until you explicitly remove the QwEventDict files.  But apart from removing the error messages no functionality was changed.
(*) QwClassLinkDef.h is not a dependency of QwClassDict.* in .auxDepends,
so changes in the LinkDef file do not trigger a regeneration of the Dict.
This would require a change in line 655 of the GNUmakefile, but I can't
figure out how to get it to do this right.  We could modify what gets
written to .auxLinkDefFiles to include the Dict target; that could become
the first line of the .auxDepends section.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1075 b9700a98-9937-0410-b777-f318d6eb5bff