Merged the wdconinc.running_sum branch into the trunk as of revision 1076, using: svn merge https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1076     https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.running_sum@1076 .
This implements a more numerically robust algorithm for calculating
the mean and RMS, and allows us to create multiple running sum
variables which may have different time ranges of validity (a burst
running sum versus the full runlet running sum, for example).
The multiple running sums will be implemented in a later revision.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1077 b9700a98-9937-0410-b777-f318d6eb5bff