Removed the implementation file for the QwSubsystemFactory (now inline), changed some names.  An error message is now returned when you try to create a subsystem that is not registered (but then it still crashes).
So, how does this work?  There is a templated concrete subsystem
factory class which allows to instantiate subsystem factories for all
subsystem types.  These factories then know how to instantiate the
subsystems.  Every subsystem class instantiates a static concrete factory
object, and in the constructor of the factory the type is associated
with the subsystem name in a static map.  On startup (before we enter
main) we thus have the map from the name to the concrete factory objects,
which allows us to create subsystems by picking the right factory from the
map.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1092 b9700a98-9937-0410-b777-f318d6eb5bff