Added a command-line flag to set the database access, "QwDatabase.accesslevel".  The options are: "off", to avoid making any database calls, "ro" for read only access, and "rw", for both read and write access.
It may be that the read-only mode doesn't really work for us
given the way that we create new runlet IDs if we can't find one
that matches our run, but we'll have to evaluate that as we continue
onward.

Two test functions ,AllowsReadAccess and AllowsWriteAccess, let
outside systems check for the ability to interact with the database
and should be called from the main loop before entering a major
database accessor function.

Shifted the options decoding out of the "ValidateConnection" into
a "ProcessOptions" function; also added a constructor that takes
the options object to set the internal parameters at construction.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1094 b9700a98-9937-0410-b777-f318d6eb5bff