Added subsystem factory support to the tracking detectors, they are now all instantiated using the factory in the qwsimtracking executable.
Implemented the geometry reader for the main detector and trigger 
scintillators (in tracking mode).  Horrible duplication of code
while awaiting more structures...  The addition of these two
systems in the code requires a change in QwTreeEventBuffer.

We are considering a strategy where you specify a parameter file
with the subsystems to be included.  Something like

  QwMainDetector MD: qweak_main_detector.map qweak_main_detector.geo
  QwDriftChamberHDC HDC: qweak_hdc.map qweak_hdc.geo
  ...



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1097 b9700a98-9937-0410-b777-f318d6eb5bff