Note: this is branch pking.burst_analysis.
This finalizes the burst analysis framework (up to trivialities, like
how to call the running sum of bursts sums to avoid confusion).

Before this branch, the running sum functionality ensured that at the
end of a run(let) we could calculate trivially the run-averaged pattern
yield, pattern difference and pattern asymmetry.  Because a pattern
is too short for some low-rate detectors, and a run(let) is too long
for the needed granularity in time, an intermediate layer with easily
variable length was introduced: the burst.

For every burst (of for example 100 helicity windows, or 25 patterns)
the sum of pattern differences and pattern yields is calculated.  At
the end of a burst, these two values are then used to calculate an
asymmetry with a properly defined and normal variance.  The calculated
asymmetries for all bursts in a run(let) are averaged at the end of a
run(let).

This framework of burst analysis is now in place.  It will work 'out of
the box' for VQWK channels, but for other modules you might have to
modify the operator+= and Ratio methods a little bit to get the variance
right (if you need that).

You can test this branch by running bin/qwmockdataanalysis.  After every
burst of 25 patterns it will print the average pattern yield and
difference, and the burst asymmetry.  At the end of the runlet it will
print the average pattern yield, difference and asymmetry, as well as
the average burst asymmetry.  There are small differences between the
average of the burst asymmetries and the average of the pattern asymmetries
but that is to be expected.  The run-averaged burst yield is of course
identical to the run-averaged pattern yield; same holds for difference.
This redundancy will be kept for diagnostics, but will be removed/disabled
for production.

Everything is weighted appropriately, so a burst with only 10 valid patterns
is weighted less than a burst with all 25 valid pattern (this can happen
because of event cuts).

The unrelated change in QwLog fixes some coloring (and other) issues in
QwLog lines that are closed with a "\n".  Or something like that...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.burst_analysis@1108 b9700a98-9937-0410-b777-f318d6eb5bff