Some inlined methods in QwGEMCluster were moved to an implementation file that also holds the ClassImp() ROOT macros.
The QwTreeEventBuffer now generates 'hits' in the main detector.  The
main detector is considered a detector with two y-separated elements
(the PMTs) with a light yield distribution dependence on x position
only (quadratic).  The light yield info is stored in fRawTime, for lack
of a better place at this time.  Access to these hits from the QwHitContainer
is not as easy as it could be, because they are still generated in region 3,
instead of their own region (which is foreseen).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1111 b9700a98-9937-0410-b777-f318d6eb5bff