Support for parametrized subsystem array filling (using map file)
All subsystems in Parity and Tracking (and the Scanner) are now
registered with subsystem factories.  The subsystem array was
modified to allow for reading in of a single file with a list of
subsystems to create, with a possibility to specify a channel
map, a set of input parameters and a geometry definition.

For an example, see QwMockDataGenerator and QwSimTracking.
There isn't really much to see there since this involved the
removal of many lines of code...  This should make it easier
to merge the QwAnalysis_* analyzers.  The filename needs to
become an options that can be specified (QwOptions), but now
it is just a hard-coded "detectors.map" that is stored in the
prminput directory.  Therefore, if you have a local copy in
your scratch directory, it will have preference over the global
copy in the QwAnalysis tree.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1122 b9700a98-9937-0410-b777-f318d6eb5bff