Changes to the blinder to have it do the blinding on the differences instead of on the asymmetry.  This will allow us to keep putting the differences and helicity into the tree and to blind the burst level quantities without any risk of leaking info.  The code allows for a combination of multipliative and additive blinding factors:
  diff_blinded = (diff_raw + A_blind * yield_raw) * f_blind

The factors A_blind and f_blind are generated within specified ranges
from a string using various degrees of magic.  Unblinding of some numbers is
(still) tested every time the blinder runs to make sure that we can unblind
the data later.

The signals that need to be blinded (only in the main detector?) should
implement the method Blind(const QwBlinder*, const VQwSubsystemParity*)
appropriately.  The second argument is the yield, but this method just
needs to cascade down to the appropriate modules.  At this point all MD
data streams (QwIntegrationPMT and QwCombinedPMT) are blinded.

The asymmetry offset is small enough that you won't notice it, but the
scaling of the differences (only!) is in the order of 10%.  If anyone is
bothered by this we can reduce the allowed scaling range.

The blinder is owned by the QwHelicityPattern class.  Interfacing should
happen through the database, but that code hasn't been changed to the new
DB schema yet (and just uses string construction of SQL commands).

Also:
- changed svn props on dictionary directories so spurious header files
  cannot play hide-and-seek there anymore.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1135 b9700a98-9937-0410-b777-f318d6eb5bff