Changes to the blinder.  Multiplicative blinding is disabled now.
To make things clear from a programmer's perspective: it was trivial
to write the code for multiplicative and/or additive blinding, and it
was easy to write it for blinding asymmetries or differences.  Even
if we still need to decide what to use, at least we have all options
available...  People who feel more strongly about blinding should
fight^H^H^H^H^H determine what to use.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1137 b9700a98-9937-0410-b777-f318d6eb5bff