Working on bring more database functionality to QwGUIView.  I was getting segfaults when choosing to "Open Database".  Appeared to be linked to fact that the QwOptions class was being used to parse database options but the command line was never being parsed.  Was getting funny interactions when I added command line parsing while preserving Michael's original command line parsing code so I reworked the command line options to use QwOptions.  (Now I no longer get seg faults when trying to "Open Database" but I do get them after interacting with the database dialog box.  Work on that later.)
I changed the behavior of QwOptions.cc so that it no longer prints the usage message and exits if there are no command line options present.  This makes using QwGUIView more sensible (no need for command line options in default mode) while not significantly breaking QwAnalysis functionality.  (QwAnalysis bombs out when it tries to replay run 0, the default option for the run command line flag.)

Note:  These changes are in the spayde.db.01_00_0000 branch, not trunk.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/spayde.db.01_00_0000@1140 b9700a98-9937-0410-b777-f318d6eb5bff