Some updates to how the blinder writes to the database.
The database has to be specified as an argument to FillDB, no pointer
is stored internally to QwBlinder.  The two functions that wrote test
values and checksum to the database have been combined and now use the
interface provided by QwDatabase instead of raw SQL commands.  Some
SQL still remains (and the old functions haven't been removed).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1141 b9700a98-9937-0410-b777-f318d6eb5bff