Note: this is the branch pking.burst_analysis.
Merge from trunk into branch over revisions 1077:1150.

Some conflicts had to be resolved (as was expected due to
interrelated changes to blinder and burst sum logic).
Merge was verified against previous branch revision.  One
regression was found: lumi detector running sum event counter
is now double of what it should be for pattern yields...
but that is not a blocker.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.burst_analysis@1151 b9700a98-9937-0410-b777-f318d6eb5bff