Added a flag to the QwMainCerenkovDetector class to enable normalization by the beam charge ("--QwMainCerenkovDetector.normalize").
When we are normalizing the MD data, we now test for "too small"
values of the beam current and don't normalize for that case.
Eventually this should be a settable parameter, but for now
I just hardcoded it as 1e-9.
The main detector normalization did not have a test for zero beam
current, and so when the beam current is  close to zero, we got
NaN values for the normalized values (and those don't get written
to the TTree).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1158 b9700a98-9937-0410-b777-f318d6eb5bff