1) Small changes in the factory structure.  Instead of having global inlined function with static variables, there is a clearer sense of ownership now. The list of registered subsystems is kept in the virtual VQwSubsystemFactory and the factory functions are static in that class.  Therefore syntax change to VQwSubsystemFactory::GetSubsystemFactory(type)->Create(name).
2) Change in name from GetSubsystem(name) to GetSubsystemByName(name) to allow
for GetSubsystemByType(type).  Both take a string as argument, so regular over-
loading was not possible.  It is preferable to use GetSubsystemByType (which
returns a vector to allow for multiple subsystems) because the names are now
not fixed at compile time and can easily change.  GetSubsystemByType is not
really functional yet, therefore not in use.  typeid is bugging me.  This will
be first used in QwHelicityPattern to get the QwHelicity subsystem.  Because
this falls outside the event loop there is no easy way of using the intersubsystem
communication framework.

3) Moved the static DefineOptions calls from the main() to the QwOptionsParity
header as calls to DefineOptions.  This removes the need to do anything subsystem
specific in the main and keeps all options related code in the respective headers.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1168 b9700a98-9937-0410-b777-f318d6eb5bff