Added support for ini-style parameter files in QwParameterFile, e.g.   [section]   key = value This is used in the detectors definition file as follows:   [QwBeamLine]   name = Injector Beamline   map = qweak_beamline.map   geom = qweak_beamline_geometry.map and   [QwDriftChamberVDC]   name = VDC   map = qweak_channel_map_vdc.map   ... etc
QwParameterFile is not limited to files anymore.  When a file is specified,
it uses the file (through a stringstream), but you can also retrieve sections
above as separate QwParameterFile objects.

Only the QwSimTracking executable uses this scheme, but this will make combining
the parity executables easier (the factory and parametrized instantiation, I mean).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1173 b9700a98-9937-0410-b777-f318d6eb5bff