Fairly complete wrapper around the ROOT file or memory-mapped file, depending on the command line flag --online (--realtime?).  It is not in use because I couldn't test the memory-mapped functionality.
To have a single executable for many conditions it is easier if we
move all the low-level ROOT interaction into a wrapper object.  This
wrapper object opens the file (be it a regular TFile or a TMapFile),
creates the tree and histogram structure (which is slighly different
for TFile and TMapFile) and takes care of output flags (--disable-trees
or --disable-histos, or --disable-mps or --disable-hel).

What previously took a screenful of code has now been reduced to

      rootfile = new QwRootFile(QwEvt.GetRunLabel());

      rootfile->ConstructHistograms(QwDetectors);
      rootfile->ConstructHistograms(QwHelPat);

      rootfile->ConstructTreeBranches(QwDetectors);
      rootfile->ConstructTreeBranches(QwHelPat);

Depending on presence of the command line option --online the first
line creates a TFile or TMapFile.

Prescaling of histogram filling (fEvents2Skip and fEvents2Save) has
not been adapted to this wrapper yet, but that should also end up in
here.  I just haven't found an elegant way to access the event number
yet, because inside the QwRootFile there is no access to the event
buffer.

Other changes:
- two small changes to QwMapFile to have identical calling behavior
as TFile.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1175 b9700a98-9937-0410-b777-f318d6eb5bff