Changes to the ROOT file wrapper.  Apparently there was again some confusion about what online and realtime mean exactly.  The flag '--online' cannot be used for the realtime display :-\
Suggestion for the use of the word 'online': connects directly to
the CODA ET stream and does not read a CODA file.

Suggestion for the use of the word 'realtime': only to be used in
connection with the QwRealTimeGUI.  QwAnalysis_Online falls in this
category...


Change to the scanner: making directories in the ROOT file is not
supported for memmap files.  This happens in a location that is not
wrapped by the QwRootFile (TDirectory* is passed and TDirectory::mkdir
is called).  For now the 'scanner' directory is disabled.


QwAnalysis_Online still crashes on QwEventRing destruction.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1182 b9700a98-9937-0410-b777-f318d6eb5bff