Updated measurement_type_ids used in ::FillDB routines.  Removed usage of QwDatabase::GetMeasurementID to retrieve appropriate measurement_type_id in these routines.  The measurement_type_ids are hard-coded in and should be stable across database versions.
Allowed measurement_type_ids are a, a12, aeo, d, d12, deo, p, r, y, ya, ye, ym, yp, and yq.  Anything with an "a" prefix is an asymmetry, anything with a "d" is a difference, and "y" is a yield.  The "12" or "eo" suffixes correspond to non-helicity correlated asymmetries or differences, either first half-second half or even-odd (in QRT member number), otherwise it is a helicity-correlated quantity.  "y" should be used for detector or lumi yields, "yp" for BPM yields, "yq" for BCM yields.  "ym", "ya", and "ye" are not currently in use and correspond to slope, angle, and energy yields.  There may be a need to add additional "a" and "d" measurement types for BPMs/BCMs to support different units for those measurements.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1188 b9700a98-9937-0410-b777-f318d6eb5bff