Modified GUI to correctly compute MD FFT and a host of other small changes ...
Still trying to figure out what is going on with the 
database access ...

There is some kind of binning / limits error in the
FFT.C macro that I am working on fixing. This error
causes the spectrum to change (dramatically) and
in weird ways, when different event cuts are used.
The GUI FFT no longer does this, and this was fixed
by paying close attention to the hostogramm binning 
and filling.

Michael





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1194 b9700a98-9937-0410-b777-f318d6eb5bff