Updates to fix database access from QwGUI.  Added new QwDatabase::ProcessOptions() member function that takes database options as parameters rather than as QwOptions&.  Updated QwGUIDatabaseContainer.cc to reflect this.  This allows QwGUI to access database without seg fault.
Had to resolve conflicts in QwGUIDatabaseContainer.cc.  Changes to that file could probably be rolled back without loss of functionality.

After updating to most recent commit, QwGUI no longer compiles for me.  It complains about missing QwGUIFFTWindowSelectionDialog.h at compile time.  This file is not present in the most recent commit.  Is it a file that did not get added to the repository correctly?



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1208 b9700a98-9937-0410-b777-f318d6eb5bff