Added QwGUIDatabase subsystem to QwGUIView in order to retrieve and plot data from the database.  This adds a new tab to the GUI, Qweak Database, and will allow the user to select run range, detector, measurement type, etc.  It may ultimately make sense to break this functionality out into existing subsystems but I have not quite wrapped my head around how to do that yet.
None of the text fields or buttons on this new tab work yet.  A single query is run and plot made when the database is open:  the asymmetry in the combined main detector as a function of run number.  The next step is to wire up the text fields and buttons so the user can plot whatever data they choose.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1227 b9700a98-9937-0410-b777-f318d6eb5bff