Created a new class called the VQwBPM: a virutal bpm for the QwBPMStripline, QwCombinedBPM and the cavity monitors. QwBPMStripline and QwCombinedBPM now derive from this class.
Created the QwEnergyCalculator which derives from the VQwDataElement.It uses any bpm or a combinedbpm to get the information that is needed for the calculation of beam energy using the equation Nuruzzaman presented in the doccument (qweak doc 1149). The devices are declared in the qweak_beamline.map. A description about the proper declaration is given in the qweak_beamline.map. Note that instead of the transport matrix elements we are using ratios of the matrix elements (see qweak_beamline.map file).

Modified the Offset(offset) funtion in the QwVQWK_Channel to be AddChannelOffset(offset). This makes is more clearer as to where the offset is being added to. So now the offset gets add to the HardwareSum and offset/ fBlocksPerEvent gets added to Blocks. When using the old Offset function to add the position offsets to the bpm readings I had to do something like Offset(position offset/4) . As far as I know, the only classes which are using this function is the beamline classes. 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1235 b9700a98-9937-0410-b777-f318d6eb5bff