Fixed a problem on the QwHelicityPattern class that was causing the pattern number checks in QwHelicity::+= operator to fail. It was related to the fact that we were using the current helicity object to get pattern information for asymmetry calculations that were being performed using references of the helicity object that was being passed in by the event ring. Wouter suggested a fix and it worked. I made those changes in here.
Also I had to remove a check that was being done on fWord[i] in the same += operator function. Checking the whole fWord[i] containing input_regsiter and mps_counter when calculating asymmetries doesn't make sense. We can only compare the pattern number and the pattern polarities within the same patterns to see if we are using events from the same pattern. Since we were comparing the input_register and the mps_counter, all the patterns was being marked off as bad patterns(-9999 ).

Now its fixed.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1309 b9700a98-9937-0410-b777-f318d6eb5bff