The flat tree is implemented using the only leaves. Tree Trimming Definition File is added that only devices on this file have tree leaves added.
In Real time GUI,
A separate tab is added in the Real Time GUI for correlation plots. In the tab, tree leaves are loaded in to two combo boxes. There is a list box to select event or pattern tree. To plot a correlation select X and Y from the combo boxes and click "Plot". 
There is also a option to set cuts on X and Y parameters. One can either give two simple cuts on two axis or give a complex expression to be pass as the cut. 

Updated real time GUI Injector tab to see summary plots for BPM position differences.

See Wiki 
	https://qweak.jlab.org/wiki/index.php/Online_Monitoring for a summary.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1323 b9700a98-9937-0410-b777-f318d6eb5bff