This commit enables the readout of the halo monitors.  The database and averages need more work before they can be uncommented but the halos are now being decoded...
Before Paul's fix every channel in a roc had to be read out to get to a later channel.  The beamline map I'm uploading is not using the fix.  I'm going to look at Paul's fix to so we don't have to call every scaler channel in the map file.  I think Buddhini is looking at this too.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1341 b9700a98-9937-0410-b777-f318d6eb5bff