Modified the QwBeamLine such that it is able to use the types and functions in QwTypes. I moved the EBeamInstrumentType enumeration in QwBeamLine.h to QwTypes.h as EQwBeamInstrumentType and added functions that are required to get device types and names using the EQwPMTInstrumentType as an example. These were then used in the LoadChannelMap() and QwBeamDetectoID in QwBeamLine to identify and sort device types.
Added a new type of PMT called a QwScalerPMT in to  EQwPMTInstrumentType in QwTypes. This being used by QwHalo for its scalers. 
Added methods for the QwHalo to decode scaler channels and write data into the tree. 
Updated the qweak_lumi.map by adding the scalers and changing the device names to match the usual notaion "qwk_****".




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1350 b9700a98-9937-0410-b777-f318d6eb5bff