Added two command line switches for input coda files: codafile-stem and codafile-ext. Everyone seems to be writing to different prefixes anyway...
Also added an end-of-burst function into the QwEventBuffer for the burst analysis.
Although most of the burst analysis happens in the QwHelicityPattern, that class has
only access to the event number in a convoluted way.  Since we don't need the exact
event number, we also don't need to worry about the event ring here, so the event buffer
is a reasonable place to put this.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1355 b9700a98-9937-0410-b777-f318d6eb5bff