Small change to the "Processing event ###" line to prevent printing out too many zero events (usually associated with EPICS events, I think).
Also, changed stopwatch to fStopwatch since it's a class member.

PS: Current trunk version of qwanalysis segfaults on startup in routines
related to QwBeamLine and event cuts...  Will look into this, but stick to
r1427 if you want a working version.  It also complains a bit when
initializing the injector beamline.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1430 b9700a98-9937-0410-b777-f318d6eb5bff