Implemented a mechanism to specify which event types a subsystem is interested in. This is relevant for a couple of people (who might not realize that yet), but mostly for the combination of tracking and parity analysis (inclusion of lumi scalers in the tracking data stream, helicity windows in tracking mode) and for the Moller analysis.
In the main detector map, the line 'mask = 0x1234' defines the event types that
the subsystem should act on.  There are 16 possible event types, represented by
the 16 bits.  The bits of all subsystems in an array are combined to determine a
bit pattern that the array uses to figure out whether to pass on the event to
subsystems.  A default mask is set for all tracking (0xfffe) and parity (0x1)
subsystems, so you most likely don't need to worry about this at all.  As usual
the scanner is messing this up, so that is the only subsystem that has a mask
defined and set to 0x0 (otherwise it ends up being 0xffff and then there is no
point (yet)).  Right now the event type is not passed to the subsystem array for
use in determining which subsystem to pass the data to, so this only works on a
subsystem-by-subsystem level.  All subsystems start empty with mask 0x0.

The tracking analyzer was changed to include a parity subsystem array.  Using the
masks above, the parity events (event type 0x1) are passed only to the parity
subsystem array, and tracking events (all other event types) are passed only to the
tracking subsystem array.  This makes the helicity subsystem happy, it does not
lose sync with every tracking event.  The helicity subsystem was added to the
detector map, but I left it as fake helicity because I didn't know which data file
now has a valid helicity data stream.

A flag HasDataLoaded() was added to the subsystem array to prevent processing when
there is nothing to process.  At a lower level (in the subsystems) this flag has
always existed and will be enforced in the near future.  If you stop seeing anything
in histograms, it is because you should call SetDataLoaded(kTRUE) after you have
finished processing your events in ProcessEvent.

Also, changed a Long_t into oolong tea in VQWK channel because ROOT doesn't follow
its own type definitions: Long_t is 4 bytes, not 8 bytes, and an 0xffffffff will
therefore overflow a signed Long_t but not an ULong_t.  The parity analyzer is
now somewhat validated on 64-bit systems (identical output as 32-bit, i.e. possibly
equally wrong...).

The histogram helper does not clear its parameter definitions at the beginning of
a read, which can lead to confusion.  You should try not to be too confused.

Disabled EPICS output completely with the flag that Han introduced.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1435 b9700a98-9937-0410-b777-f318d6eb5bff